// client/src/types/project.ts
export interface ProjectProcessStep {
    number: string;
    title: string;
    description: string;
}


/* ============================================================
   PROJECT SCREENSHOT
============================================================ */

export interface ProjectScreenshot {
    id: string;
    image: string;
    title?: string;
    caption?: string;
}


/* ============================================================
   PROJECT TECHNOLOGY
============================================================ */

export interface ProjectTechnology {
    name: string;
    category?: string;
    icon?: string;
}


/* ============================================================
   PROJECT FEATURE
============================================================ */

export interface ProjectFeature {
    number: string;
    title: string;
    description: string;
}


/* ============================================================
   PROJECT RESULT
============================================================ */

export interface ProjectResult {
    value: string;
    label: string;
    description?: string;
}


/* ============================================================
   PROJECT CREDIT
============================================================ */

export interface ProjectCredit {
    role: string;
    name: string;
}


/* ============================================================
   MAIN PROJECT
============================================================ */

export interface Project {

    /* --------------------------------------------------------
       BASIC INFORMATION
    -------------------------------------------------------- */

    id: string;

    title: string;

    slug: string;

    description: string;

    category: string;

    year: number;


    /* --------------------------------------------------------
       PROJECT META
    -------------------------------------------------------- */

    client?: string;

    role?: string;

    duration?: string;


    /* --------------------------------------------------------
       HERO
    -------------------------------------------------------- */

    coverImage: string;


    /* --------------------------------------------------------
       EXTERNAL LINKS
    -------------------------------------------------------- */

    liveUrl?: string;

    githubUrl?: string;


    /* --------------------------------------------------------
       CASE STUDY
    -------------------------------------------------------- */

    overview?: string;

    challenge?: string;


    /* --------------------------------------------------------
       PROCESS
    -------------------------------------------------------- */

    process?: ProjectProcessStep[];


    /* --------------------------------------------------------
       SCREENSHOTS
    -------------------------------------------------------- */

    screenshots: ProjectScreenshot[];


    /* --------------------------------------------------------
       TECHNOLOGIES
    -------------------------------------------------------- */

    technologies: ProjectTechnology[];


    /* --------------------------------------------------------
       FEATURES
    -------------------------------------------------------- */

    features?: ProjectFeature[];


    /* --------------------------------------------------------
       RESULTS
    -------------------------------------------------------- */

    results?: ProjectResult[];


    /* --------------------------------------------------------
       CREDITS
    -------------------------------------------------------- */

    credits?: ProjectCredit[];


    /* --------------------------------------------------------
       PUBLISHING
    -------------------------------------------------------- */

    published: boolean;
}